:recycle: refactor(typecheck): 迁移类型检查模块到 inference 目录
将原 checking/ 目录的代码迁移到 inference/ 目录，统一类型推断模块：
- 将 assignment、bounds、compatibility、subtyping 移入 inference/
- 删除 specialize.rs，相关功能整合到推断流程中
- 更新 inference/mod.rs 导入新模块
- 移除过时的测试文件 generic.rs 并更新相关测试
