+++ src/frontend/typecheck/inference/expressions.rs
@@ -2,7 +2,8 @@
 
 //! 表达式类型推断
 //!
-//! 实现各种表达式的类型推断
+//! 实现各种表达式的类型推断。
+//! 使用统一的 ScopeManager 管理变量作用域。
 
 use crate::util::diagnostic::{ErrorCodeDefinition, Result};
 use crate::frontend::core::parser::ast::{BinOp, UnOp};
@@ -10,11 +11,17 @@ use crate::frontend::core::type_system::{MonoType, PolyType, TypeConstraintSolve
 use crate::frontend::typecheck::overload;
 use std::collections::HashMap;
 
+use super::scope::ScopeManager;
+
 /// 表达式类型推断器
-pub struct ExprInferrer<'a> {
+///
+/// 使用统一的 ScopeManager 管理变量作用域，
+/// 不再维护独立的作用域栈。
+pub struct ExpressionInferrer<'a> {
+    /// 共享的作用域管理器
+    scope: &'a mut ScopeManager,
+    /// 约束求解器
     solver: &'a mut TypeConstraintSolver,
-    /// 变量环境栈：每一层是一个作用域
-    scopes: Vec<HashMap<String, PolyType>>,
     /// 当前活跃的循环标签
     loop_labels: Vec<String>,
     /// 重载候选存储引用
@@ -23,18 +30,18 @@ pub struct ExprInferrer<'a> {
     native_signatures: &'a HashMap<String, MonoType>,
 }
 
-impl<'a> ExprInferrer<'a> {
+impl<'a> ExpressionInferrer<'a> {
     /// 创建新的表达式推断器
     pub fn new(
+        scope: &'a mut ScopeManager,
         solver: &'a mut TypeConstraintSolver,
         overload_candidates: &'a HashMap<String, Vec<overload::OverloadCandidate>>,
     ) -> Self {
-        // 默认空 native_signatures，使用 with_native_signatures 来传入实际签名
         static EMPTY_NATIVE: std::sync::LazyLock<HashMap<String, MonoType>> =
             std::sync::LazyLock::new(HashMap::new);
         Self {
+            scope,
             solver,
-            scopes: vec![HashMap::new()], // Global scope
             loop_labels: Vec::new(),
             overload_candidates,
             native_signatures: &EMPTY_NATIVE,
@@ -43,13 +50,14 @@ impl<'a> ExprInferrer<'a> {
 
     /// 创建带 native 函数签名的表达式推断器
     pub fn with_native_signatures(
+        scope: &'a mut ScopeManager,
         solver: &'a mut TypeConstraintSolver,
         overload_candidates: &'a HashMap<String, Vec<overload::OverloadCandidate>>,
         native_signatures: &'a HashMap<String, MonoType>,
     ) -> Self {
         Self {
+            scope,
             solver,
-            scopes: vec![HashMap::new()],
             loop_labels: Vec::new(),
             overload_candidates,
             native_signatures,
@@ -61,13 +69,13 @@ impl<'a> ExprInferrer<'a> {
         self.solver
     }
 
-    /// 添加变量到当前作用域（无遮蔽检查）
+    /// 添加变量到当前作用域
     pub fn add_var(
         &mut self,
         name: String,
         poly: PolyType,
     ) {
-        self.scopes.last_mut().unwrap().insert(name, poly);
+        self.scope.add_var(name, poly);
     }
 
     /// 检查变量是否存在于任何作用域中
@@ -75,29 +83,22 @@ impl<'a> ExprInferrer<'a> {
         &self,
         name: &str,
     ) -> bool {
-        for scope in &self.scopes {
-            if scope.contains_key(name) {
-                return true;
-            }
-        }
-        false
+        self.scope.var_in_any_scope(name)
     }
 
     /// 尝试添加变量到当前作用域（带遮蔽检查）
-    /// 如果变量已存在于任何作用域中，返回错误（遮蔽）
     pub fn try_add_var(
         &mut self,
         name: String,
         poly: PolyType,
         span: crate::util::span::Span,
     ) -> Result<()> {
-        // 检查是否在所有作用域中存在同名变量（遮蔽检查）
-        if self.var_exists_in_any_scope(&name) {
+        if self.scope.var_in_any_scope(&name) {
             return Err(ErrorCodeDefinition::variable_shadowing(&name)
                 .at(span)
                 .build());
         }
-        self.scopes.last_mut().unwrap().insert(name, poly);
+        self.scope.add_var(name, poly);
         Ok(())
     }
 
@@ -106,11 +107,7 @@ impl<'a> ExprInferrer<'a> {
         &self,
         name: &str,
     ) -> bool {
-        if let Some(current_scope) = self.scopes.last() {
-            current_scope.contains_key(name)
-        } else {
-            false
-        }
+        self.scope.var_in_current_scope(name)
     }
 
     /// 获取变量（从最内层作用域开始查找）
@@ -118,45 +115,30 @@ impl<'a> ExprInferrer<'a> {
         &self,
         name: &str,
     ) -> Option<&PolyType> {
-        for scope in self.scopes.iter().rev() {
-            if let Some(poly) = scope.get(name) {
-                return Some(poly);
-            }
-        }
-        None
+        self.scope.get_var(name)
     }
 
     /// 获取所有变量（从所有作用域）
     pub fn get_all_vars(&self) -> HashMap<String, PolyType> {
-        let mut result = HashMap::new();
-        for scope in &self.scopes {
-            for (name, poly) in scope {
-                if !result.contains_key(name) {
-                    result.insert(name.clone(), poly.clone());
-                }
-            }
-        }
-        result
+        self.scope.vars()
     }
 
     /// 进入新的作用域
     pub fn enter_scope(&mut self) {
-        self.scopes.push(HashMap::new());
+        self.scope.enter_scope();
     }
 
     /// 退出当前作用域
     pub fn exit_scope(&mut self) {
-        if self.scopes.len() > 1 {
-            self.scopes.pop();
-        }
+        self.scope.exit_scope();
     }
 
     /// 获取当前作用域层级
     pub fn scope_level(&self) -> usize {
-        self.scopes.len()
+        self.scope.scope_level()
     }
 
-    /// 进入循环并注册标签（如果有）
+    /// 进入循环并注册标签
     pub fn enter_loop(
         &mut self,
         label: Option<&str>,
@@ -166,7 +148,7 @@ impl<'a> ExprInferrer<'a> {
         }
     }
 
-    /// 退出循环并移除标签（如果有）
+    /// 退出循环并移除标签
     pub fn exit_loop(
         &mut self,
         label: Option<&str>,
@@ -219,30 +201,25 @@ impl<'a> ExprInferrer<'a> {
                 } else if let (MonoType::List(left_elem), MonoType::List(right_elem)) =
                     (left, right)
                 {
-                    // 列表拼接：统一元素类型，返回统一后的列表类型
                     let _ = self.solver.unify(left_elem, right_elem);
                     let elem_ty = self.solver.resolve_type(left_elem);
                     Ok(MonoType::List(Box::new(elem_ty)))
                 } else {
-                    // 对于字符串连接和其他类型，使用类型变量
                     let var = self.solver.new_var();
                     Ok(var)
                 }
             }
             BinOp::Mod => {
-                // 取模运算：两个操作数必须是相同数值类型，返回相同类型
                 if let (MonoType::Int(_), MonoType::Int(_)) = (left, right) {
                     Ok(left.clone())
                 } else if let (MonoType::Float(_), MonoType::Float(_)) = (left, right) {
                     Ok(left.clone())
                 } else {
-                    // 尝试统一两个类型（忽略错误，返回左操作数类型）
                     let _ = self.solver.unify(left, right);
                     Ok(left.clone())
                 }
             }
             BinOp::Eq | BinOp::Neq | BinOp::Lt | BinOp::Le | BinOp::Gt | BinOp::Ge => {
-                // 比较运算：尝试统一两个操作数
                 let _ = self.solver.unify(left, right);
                 Ok(MonoType::Bool)
             }
@@ -258,21 +235,15 @@ impl<'a> ExprInferrer<'a> {
                 }
             }
             BinOp::Range => {
-                // 范围表达式的类型是列表，元素类型与操作数相同
                 let elem_ty = if left == right {
                     left.clone()
                 } else {
-                    // 尝试统一
                     let _ = self.solver.unify(left, right);
                     left.clone()
                 };
                 Ok(MonoType::List(Box::new(elem_ty)))
             }
-            BinOp::Assign => {
-                // 赋值表达式的类型是 Void
-                // 字段可变性检查在 MutChecker 中进行
-                Ok(MonoType::Void)
-            }
+            BinOp::Assign => Ok(MonoType::Void),
         }
     }
 
@@ -296,10 +267,7 @@ impl<'a> ExprInferrer<'a> {
                 }
             }
             UnOp::Deref => {
-                // 解引用：*ptr 返回内部类型
-                // ptr 应该是 *T 形式
                 if let MonoType::TypeRef(inner) = expr {
-                    // 提取内部类型（去掉 * 前缀）
                     let inner_type = inner.trim_start_matches('*').to_string();
                     Ok(MonoType::TypeRef(inner_type))
                 } else {
@@ -320,13 +288,12 @@ impl<'a> ExprInferrer<'a> {
 
             // 变量
             crate::frontend::core::parser::ast::Expr::Var(name, span) => {
-                let poly = self.get_var(name).cloned();
+                let poly = self.scope.get_var(name).cloned();
                 if let Some(poly) = poly {
                     let ty = self.solver.instantiate(&poly);
                     let resolved = self.solver.resolve_type(&ty);
                     Ok(resolved)
                 } else {
-                    // 返回错误
                     Err(ErrorCodeDefinition::unknown_variable(name)
                         .at(*span)
                         .build())
@@ -337,25 +304,19 @@ impl<'a> ExprInferrer<'a> {
             crate::frontend::core::parser::ast::Expr::BinOp {
                 op, left, right, ..
             } => {
-                // 先推断右操作数类型
                 let right_ty = self.infer_expr(right)?;
 
-                // 处理赋值操作：将左操作数（变量）的类型与右操作数类型统一
                 if matches!(op, BinOp::Assign) {
                     if let crate::frontend::core::parser::ast::Expr::Var(var_name, _) =
                         left.as_ref()
                     {
-                        // 获取变量的当前类型
-                        if let Some(poly) = self.get_var(var_name).cloned() {
-                            // 将变量类型与右操作数类型统一（忽略错误，因为类型不匹配会在其他地方报错）
+                        if let Some(poly) = self.scope.get_var(var_name).cloned() {
                             let _ = self.solver.unify(&poly.body, &right_ty);
                         }
                     }
-                    // 赋值表达式的类型是 Void
                     return Ok(MonoType::Void);
                 }
 
-                // 非赋值操作：正常推断
                 let left_ty = self.infer_expr(left)?;
                 self.infer_binary(op, &left_ty, &right_ty)
             }
@@ -375,11 +336,9 @@ impl<'a> ExprInferrer<'a> {
             // 列表
             crate::frontend::core::parser::ast::Expr::List(elems, _) => {
                 if elems.is_empty() {
-                    // 空列表：元素类型是类型变量
                     let elem_ty = self.solver.new_var();
                     Ok(MonoType::List(Box::new(elem_ty)))
                 } else {
-                    // 非空列表：推断所有元素类型并统一
                     let mut iter = elems.iter();
                     let first = iter.next().expect("non-empty list must have first element");
                     let mut elem_ty = self.infer_expr(first)?;
@@ -395,12 +354,10 @@ impl<'a> ExprInferrer<'a> {
             // 字典
             crate::frontend::core::parser::ast::Expr::Dict(pairs, _) => {
                 if pairs.is_empty() {
-                    // 空字典：键值类型都是类型变量
                     let key_ty = self.solver.new_var();
                     let value_ty = self.solver.new_var();
                     Ok(MonoType::Dict(Box::new(key_ty), Box::new(value_ty)))
                 } else {
-                    // 非空字典
                     let mut key_ty = None;
                     let mut value_ty = None;
                     for (k, v) in pairs {
@@ -431,7 +388,6 @@ impl<'a> ExprInferrer<'a> {
                     MonoType::List(elem_ty) => Ok(*elem_ty),
                     MonoType::Dict(_key_ty, value_ty) => Ok(*value_ty),
                     MonoType::Tuple(types) => {
-                        // 静态元组下标检查
                         if let crate::frontend::core::parser::ast::Expr::Lit(
                             crate::frontend::core::lexer::tokens::Literal::Int(i),
                             _,
@@ -447,14 +403,10 @@ impl<'a> ExprInferrer<'a> {
                                 .build())
                             }
                         } else {
-                            // 动态下标：返回类型变量
                             Ok(self.solver.new_var())
                         }
                     }
-                    _ => {
-                        // 动态下标：返回类型变量
-                        Ok(self.solver.new_var())
-                    }
+                    _ => Ok(self.solver.new_var()),
                 }
             }
 
@@ -462,7 +414,6 @@ impl<'a> ExprInferrer<'a> {
             crate::frontend::core::parser::ast::Expr::FieldAccess {
                 expr: obj, field, ..
             } => {
-                // 尝试构建命名空间路径（如 std.io）
                 fn extract_namespace_path(
                     expr: &crate::frontend::core::parser::ast::Expr
                 ) -> Option<String> {
@@ -479,18 +430,15 @@ impl<'a> ExprInferrer<'a> {
                     }
                 }
 
-                // 首先推断对象类型
                 let obj_ty = self.infer_expr(obj)?;
+                let obj_ty = self.solver.resolve_type(&obj_ty);
 
-                // 如果对象是命名空间（StructType 或 Fn），尝试解析完整路径
                 let namespace_path = extract_namespace_path(obj);
                 if let Some(ns_path) = namespace_path {
-                    // 尝试作为命名空间访问处理（如 std.io）
                     let full_path = format!("{}.{}", ns_path, field);
                     if let Some(sig) = self.native_signatures.get(&full_path).cloned() {
                         return Ok(sig);
                     }
-                    // 如果是子模块访问，返回虚拟函数类型让后续调用处理
                     if self
                         .native_signatures
                         .keys()
@@ -524,44 +472,35 @@ impl<'a> ExprInferrer<'a> {
 
             // 函数调用
             crate::frontend::core::parser::ast::Expr::Call { func, args, .. } => {
-                // 0. 检查是否是 Native("...") 表达式
+                // 检查 Native("...") 表达式
                 if let crate::frontend::core::parser::ast::Expr::Var(ref name, _) = **func {
                     if name == "Native" {
-                        // Native("symbol_name") 表达式
-                        // 从参数中提取 native 函数名
                         if args.len() == 1 {
                             if let crate::frontend::core::parser::ast::Expr::Lit(
                                 crate::frontend::core::lexer::tokens::Literal::String(native_name),
                                 _,
                             ) = &args[0]
                             {
-                                // 查找 native 函数签名
                                 if let Some(sig) = self.native_signatures.get(native_name).cloned()
                                 {
                                     return Ok(sig);
                                 }
-                                // 未注册的 native 函数：返回类型变量（允许用户定义的 native）
                                 return Ok(self.solver.new_var());
                             }
                         }
-                        // Native 调用参数不合法：返回类型变量
                         return Ok(self.solver.new_var());
                     }
                 }
 
-                // 1. 首先推断函数表达式
                 let func_ty = self.infer_expr(func)?;
 
-                // 2. 推断所有参数类型
                 let arg_types: Vec<MonoType> = args
                     .iter()
                     .map(|arg| self.infer_expr(arg))
                     .collect::<Result<Vec<_>, _>>()?;
 
-                // 3. 尝试重载解析
-                // 检查 func 是否为标识符
+                // 重载解析
                 if let crate::frontend::core::parser::ast::Expr::Var(ref name, _) = **func {
-                    // 尝试从重载候选解析
                     if overload::has_overloads(self.overload_candidates, name) {
                         match overload::resolve_overload_from_env(
                             self.overload_candidates,
@@ -569,38 +508,29 @@ impl<'a> ExprInferrer<'a> {
                             &arg_types,
                         ) {
                             Ok(candidate) => {
-                                // 重载解析成功，返回匹配的返回类型
                                 return Ok(candidate.return_type.clone());
                             }
                             Err(_e) => {
-                                // 重载解析失败，尝试泛型 fallback
                                 if let Some(generic_candidate) = overload::resolve_generic_fallback(
                                     self.overload_candidates,
                                     name,
                                     &arg_types,
                                 ) {
-                                    // 泛型实例化成功，返回实例化后的返回类型
                                     let return_type = overload::instantiate_return_type(
                                         generic_candidate,
                                         &arg_types,
                                     );
                                     return Ok(return_type);
                                 }
-                                // 无匹配，返回类型变量
-                                // 重载解析失败，返回类型变量（让后续类型检查捕获错误）
-                                // 这里不直接报错，因为后续可能有其他错误信息
                                 return Ok(self.solver.new_var());
                             }
                         }
                     }
                 }
-                // 4. 非重载情况或重载失败：
-                // 函数调用返回类型变量（具体类型需要在调用点确定）
-                // 如果函数类型已知，使用其返回类型
+
                 if let MonoType::Fn { return_type, .. } = func_ty {
                     return Ok(*return_type);
                 }
-                // 否则返回类型变量（具体类型在调用点确定）
                 Ok(self.solver.new_var())
             }
 
@@ -612,7 +542,6 @@ impl<'a> ExprInferrer<'a> {
                 else_branch,
                 ..
             } => {
-                // 推断条件类型（必须是 Bool）
                 let cond_ty = self.infer_expr(condition)?;
                 if cond_ty != MonoType::Bool {
                     return Err(ErrorCodeDefinition::condition_type_mismatch(&format!(
@@ -622,13 +551,11 @@ impl<'a> ExprInferrer<'a> {
                     .build());
                 }
 
-                // 推断 then 分支（独立作用域）
-                self.enter_scope();
+                self.scope.enter_scope();
                 let then_result = self.infer_block(then_branch, true, None);
-                self.exit_scope();
+                self.scope.exit_scope();
                 let _then_ty = then_result?;
 
-                // 推断 elif 分支
                 for (elif_cond, elif_block) in elif_branches {
                     let elif_cond_ty = self.infer_expr(elif_cond)?;
                     if elif_cond_ty != MonoType::Bool {
@@ -638,22 +565,18 @@ impl<'a> ExprInferrer<'a> {
                         ))
                         .build());
                     }
-                    // elif 分支使用独立作用域
-                    self.enter_scope();
+                    self.scope.enter_scope();
                     let elif_result = self.infer_block(elif_block, true, None);
-                    self.exit_scope();
+                    self.scope.exit_scope();
                     let _ = elif_result?;
                 }
 
-                // 推断 else 分支（如果有）
                 if let Some(else_block) = else_branch {
-                    // else 分支使用独立作用域
-                    self.enter_scope();
+                    self.scope.enter_scope();
                     let else_result = self.infer_block(else_block, true, None);
-                    self.exit_scope();
+                    self.scope.exit_scope();
                     else_result
                 } else {
-                    // 没有 else 分支时返回 Void
                     Ok(MonoType::Void)
                 }
             }
@@ -665,7 +588,6 @@ impl<'a> ExprInferrer<'a> {
                 label,
                 ..
             } => {
-                // 推断条件类型
                 let cond_ty = self.infer_expr(condition)?;
                 if cond_ty != MonoType::Bool {
                     return Err(ErrorCodeDefinition::condition_type_mismatch(&format!(
@@ -675,63 +597,46 @@ impl<'a> ExprInferrer<'a> {
                     .build());
                 }
 
-                // 注册循环标签
                 self.enter_loop(label.as_deref());
 
-                // While 循环体使用独立作用域
-                self.enter_scope();
+                self.scope.enter_scope();
                 let result = self.infer_block(body, true, None);
-                self.exit_scope();
+                self.scope.exit_scope();
 
-                // 移除循环标签
                 self.exit_loop(label.as_deref());
 
                 result?;
-                // While 表达式总是返回 Void
                 Ok(MonoType::Void)
             }
 
             // For 循环
             crate::frontend::core::parser::ast::Expr::For {
                 var,
-                var_mut: _, // TODO: 使用 var_mut 进行类型推断
+                var_mut: _,
                 iterable,
                 body,
                 label,
                 span,
             } => {
-                // 推断可迭代对象类型
                 let iter_ty = self.infer_expr(iterable)?;
 
-                // 根据可迭代对象类型推导循环变量类型
                 let element_type = match &iter_ty {
-                    // List<T> -> T
                     MonoType::List(elem_ty) => *elem_ty.clone(),
-                    // Tuple 类型：根据位置推导（这里简化处理，使用类型变量）
-                    MonoType::Tuple(_elems) => {
-                        // TODO: 支持 Tuple 的元素类型推导
-                        self.solver.new_var()
-                    }
-                    // Dict 类型：键值对类型（这里简化处理，使用类型变量）
+                    MonoType::Tuple(_elems) => self.solver.new_var(),
                     MonoType::Dict(key_ty, value_ty) => {
-                        // Dict 遍历时，元素是 (Key, Value) 元组
                         MonoType::Tuple(vec![*key_ty.clone(), *value_ty.clone()])
                     }
-                    // 其他类型：使用类型变量
                     _ => self.solver.new_var(),
                 };
 
-                // 注册循环标签
                 self.enter_loop(label.as_deref());
 
-                // 进入循环体作用域，添加迭代变量（带遮蔽检查）
-                self.enter_scope();
+                self.scope.enter_scope();
                 let result = self
                     .try_add_var(var.clone(), PolyType::mono(element_type), *span)
                     .and_then(|_| self.infer_block(body, true, None));
 
-                self.exit_scope();
-                // 移除循环标签
+                self.scope.exit_scope();
                 self.exit_loop(label.as_deref());
                 result
             }
@@ -746,8 +651,6 @@ impl<'a> ExprInferrer<'a> {
 
             // Break 表达式
             crate::frontend::core::parser::ast::Expr::Break(label, _) => {
-                // Break 表达式总是返回 Void
-                // 如果有标签，需要验证标签是否有效
                 if let Some(l) = label {
                     if !self.has_label(l) {
                         return Err(ErrorCodeDefinition::unknown_label(l).build());
@@ -758,8 +661,6 @@ impl<'a> ExprInferrer<'a> {
 
             // Continue 表达式
             crate::frontend::core::parser::ast::Expr::Continue(label, _) => {
-                // Continue 表达式总是返回 Void
-                // 如果有标签，需要验证标签是否有效
                 if let Some(l) = label {
                     if !self.has_label(l) {
                         return Err(ErrorCodeDefinition::unknown_label(l).build());
@@ -773,7 +674,6 @@ impl<'a> ExprInferrer<'a> {
                 expr, target_type, ..
             } => {
                 let _ = self.infer_expr(expr)?;
-                // 将 AST Type 转换为 MonoType
                 let target_mono: MonoType = target_type.clone().into();
                 Ok(target_mono)
             }
@@ -791,58 +691,51 @@ impl<'a> ExprInferrer<'a> {
                 body,
                 ..
             } => {
-                // 为函数参数创建类型变量并添加到作用域
-                self.enter_scope();
+                self.scope.enter_scope();
                 for param in params {
                     let param_ty = self.solver.new_var();
-                    self.add_var(param.name.clone(), PolyType::mono(param_ty));
+                    self.scope
+                        .add_var(param.name.clone(), PolyType::mono(param_ty));
                 }
 
-                // 推断函数体
                 let ret_mono: MonoType = return_type.clone().map_or(MonoType::Void, |t| t.into());
                 let body_ty = self.infer_block(body, true, Some(&ret_mono))?;
 
-                // 验证返回类型
                 if return_type.is_some() {
-                    // 已经在上面计算了 ret_mono
-                    // TODO: 检查 body_ty 是否可以赋值给 ret_mono
                     let _ = self.solver.unify(&body_ty, &ret_mono);
                 }
 
-                self.exit_scope();
+                self.scope.exit_scope();
 
-                // 构建函数类型
                 let param_types: Vec<MonoType> =
                     params.iter().map(|_| self.solver.new_var()).collect();
                 let return_type_box =
                     Box::new(return_type.clone().map_or(MonoType::Void, |t| t.into()));
 
-                // 将函数类型添加到作用域
                 let fn_type = MonoType::Fn {
                     params: param_types,
                     return_type: return_type_box,
                     is_async: false,
                 };
-                self.add_var(name.clone(), PolyType::mono(fn_type.clone()));
+                self.scope
+                    .add_var(name.clone(), PolyType::mono(fn_type.clone()));
 
                 Ok(fn_type)
             }
 
             // Lambda 表达式
             crate::frontend::core::parser::ast::Expr::Lambda { params, body, .. } => {
-                // 为参数创建作用域
-                self.enter_scope();
+                self.scope.enter_scope();
                 for param in params {
                     let param_ty = self.solver.new_var();
-                    self.add_var(param.name.clone(), PolyType::mono(param_ty));
+                    self.scope
+                        .add_var(param.name.clone(), PolyType::mono(param_ty));
                 }
 
-                // 推断函数体
                 let body_ty = self.infer_block(body, true, None)?;
 
-                self.exit_scope();
+                self.scope.exit_scope();
 
-                // 构建函数类型
                 let param_types: Vec<MonoType> =
                     params.iter().map(|_| self.solver.new_var()).collect();
 
@@ -853,14 +746,13 @@ impl<'a> ExprInferrer<'a> {
                 })
             }
 
-            // Match 表达式（简化实现）
+            // Match 表达式
             crate::frontend::core::parser::ast::Expr::Match { expr, .. } => {
                 let _expr_ty = self.infer_expr(expr)?;
-                // 简化：返回类型变量
                 Ok(self.solver.new_var())
             }
 
-            // Try 表达式（简化实现）
+            // Try 表达式
             crate::frontend::core::parser::ast::Expr::Try { expr, .. } => self.infer_expr(expr),
 
             // Ref 表达式
@@ -871,7 +763,6 @@ impl<'a> ExprInferrer<'a> {
 
             // Unsafe 块
             crate::frontend::core::parser::ast::Expr::Unsafe { body, .. } => {
-                // 推断块内最后一个表达式的类型作为 unsafe 块的类型
                 if let Some(last_expr) = &body.expr {
                     self.infer_expr(last_expr)
                 } else {
@@ -887,14 +778,12 @@ impl<'a> ExprInferrer<'a> {
                 condition,
                 ..
             } => {
-                // 推断可迭代对象
                 let _iter_ty = self.infer_expr(iterable)?;
 
-                // 进入列表推导作用域
-                self.enter_scope();
-                self.add_var(var.clone(), PolyType::mono(MonoType::Char));
+                self.scope.enter_scope();
+                self.scope
+                    .add_var(var.clone(), PolyType::mono(MonoType::Char));
 
-                // 推断元素类型
                 let elem_ty = if let Some(cond) = condition {
                     let _cond_ty = self.infer_expr(cond)?;
                     self.infer_expr(element)?
@@ -902,7 +791,7 @@ impl<'a> ExprInferrer<'a> {
                     self.infer_expr(element)?
                 };
 
-                self.exit_scope();
+                self.scope.exit_scope();
 
                 Ok(MonoType::List(Box::new(elem_ty)))
             }
@@ -916,12 +805,10 @@ impl<'a> ExprInferrer<'a> {
         _allow_unit: bool,
         _expected_type: Option<&MonoType>,
     ) -> Result<MonoType> {
-        // 推断所有语句
         for stmt in &block.stmts {
             self.infer_stmt(stmt)?;
         }
 
-        // 推断最终表达式
         if let Some(expr) = &block.expr {
             self.infer_expr(expr)
         } else {
@@ -945,7 +832,6 @@ impl<'a> ExprInferrer<'a> {
                 initializer,
                 ..
             } => {
-                // 推断初始化表达式类型
                 let init_ty = if let Some(expr) = initializer {
                     self.infer_expr(expr)?
                 } else {
@@ -954,9 +840,7 @@ impl<'a> ExprInferrer<'a> {
                         .map_or_else(|| self.solver.new_var(), |t| t.clone().into())
                 };
 
-                // 如果变量已存在于任何作用域中，则是赋值操作，不进行遮蔽检查
-                // 否则，添加变量并进行遮蔽检查
-                if !self.var_exists_in_any_scope(name) {
+                if !self.scope.var_in_any_scope(name) {
                     self.try_add_var(name.clone(), PolyType::mono(init_ty), stmt.span)?;
                 }
                 Ok(())
@@ -968,7 +852,6 @@ impl<'a> ExprInferrer<'a> {
                 type_annotation,
                 ..
             } => {
-                // 创建函数类型
                 let param_types: Vec<MonoType> =
                     params.iter().map(|_| self.solver.new_var()).collect();
 
@@ -982,18 +865,15 @@ impl<'a> ExprInferrer<'a> {
                     is_async: false,
                 };
 
-                // 添加函数到作用域
-                self.add_var(name.clone(), PolyType::mono(fn_type));
+                self.scope.add_var(name.clone(), PolyType::mono(fn_type));
 
-                // 进入函数作用域
-                self.enter_scope();
+                self.scope.enter_scope();
 
-                // 添加参数到作用域
                 for (param, param_ty) in params.iter().zip(param_types.iter()) {
-                    self.add_var(param.name.clone(), PolyType::mono(param_ty.clone()));
+                    self.scope
+                        .add_var(param.name.clone(), PolyType::mono(param_ty.clone()));
                 }
 
-                // 推断函数体
                 let block = crate::frontend::core::parser::ast::Block {
                     stmts: stmts.clone(),
                     expr: expr.clone(),
@@ -1001,13 +881,13 @@ impl<'a> ExprInferrer<'a> {
                 };
                 let _ = self.infer_block(&block, true, Some(&return_type))?;
 
-                self.exit_scope();
+                self.scope.exit_scope();
                 Ok(())
             }
-            _ => {
-                // 其他语句类型暂不处理
-                Ok(())
-            }
+            _ => Ok(()),
         }
     }
 }
+
+/// 向后兼容：ExprInferrer 是 ExpressionInferrer 的类型别名
+pub type ExprInferrer<'a> = ExpressionInferrer<'a>;
