+++ src/frontend/typecheck/tests/basic.rs
@@ -5,6 +5,7 @@ use crate::frontend::core::lexer::tokens::Literal;
 use crate::frontend::core::parser::ast::{BinOp, Expr, UnOp};
 use crate::frontend::core::type_system::{MonoType, PolyType, TypeConstraintSolver, TypeBinding};
 use crate::frontend::typecheck::inference::ExprInferrer;
+use crate::frontend::typecheck::inference::ScopeManager;
 use crate::frontend::typecheck::*;
 use crate::util::span::Span;
 
@@ -13,7 +14,8 @@ use crate::util::span::Span;
 fn test_literal_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 整数
     let int_lit = Expr::Lit(Literal::Int(42), Span::default());
@@ -23,7 +25,8 @@ fn test_literal_inference() {
     // 浮点数
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
     let float_lit = Expr::Lit(Literal::Float(3.14), Span::default());
     let float_ty = inferrer.infer_expr(&float_lit).unwrap();
     assert_eq!(float_ty, MonoType::Float(64));
@@ -31,7 +34,8 @@ fn test_literal_inference() {
     // 布尔值
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
     let bool_lit = Expr::Lit(Literal::Bool(true), Span::default());
     let bool_ty = inferrer.infer_expr(&bool_lit).unwrap();
     assert_eq!(bool_ty, MonoType::Bool);
@@ -39,7 +43,8 @@ fn test_literal_inference() {
     // 字符串
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
     let str_lit = Expr::Lit(Literal::String("hello".to_string()), Span::default());
     let str_ty = inferrer.infer_expr(&str_lit).unwrap();
     assert_eq!(str_ty, MonoType::String);
@@ -51,7 +56,8 @@ fn test_binop_inference() {
     // 加法
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(1), Span::default());
     let right = Expr::Lit(Literal::Int(2), Span::default());
@@ -75,7 +81,8 @@ fn test_binop_inference() {
 fn test_comparison_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(1), Span::default());
     let right = Expr::Lit(Literal::Int(2), Span::default());
@@ -97,7 +104,8 @@ fn test_comparison_inference() {
 fn test_logical_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Bool(true), Span::default());
     let right = Expr::Lit(Literal::Bool(false), Span::default());
@@ -117,7 +125,8 @@ fn test_logical_inference() {
 fn test_tuple_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let elems = vec![
         Expr::Lit(Literal::Int(1), Span::default()),
@@ -139,7 +148,8 @@ fn test_tuple_inference() {
 fn test_list_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let elems = vec![
         Expr::Lit(Literal::Int(1), Span::default()),
@@ -199,7 +209,8 @@ fn test_type_resolution() {
 fn test_unop_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 负数
     let expr = Expr::Lit(Literal::Int(5), Span::default());
@@ -220,7 +231,8 @@ fn test_unop_inference() {
 fn test_string_concat() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::String("hello".to_string()), Span::default());
     let right = Expr::Lit(Literal::String(" world".to_string()), Span::default());
@@ -293,7 +305,8 @@ fn test_type_annotation() {
 fn test_mixed_type_list() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 混合类型列表
     let elems = vec![
@@ -314,7 +327,8 @@ fn test_mixed_type_list() {
 fn test_division_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(10), Span::default());
     let right = Expr::Lit(Literal::Int(3), Span::default());
@@ -337,7 +351,8 @@ fn test_division_inference() {
 fn test_modulo_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(10), Span::default());
     let right = Expr::Lit(Literal::Int(3), Span::default());
@@ -360,7 +375,8 @@ fn test_modulo_inference() {
 fn test_float_binop_inference() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Float(3.14), Span::default());
     let right = Expr::Lit(Literal::Float(2.0), Span::default());
@@ -381,7 +397,8 @@ fn test_float_binop_inference() {
 fn test_empty_tuple() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let tuple = Expr::Tuple(vec![], Span::default());
 
@@ -434,7 +451,8 @@ fn test_add_constraint() {
 fn test_type_inferrer_creation() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 验证推断器创建成功
     assert!(inferrer.solver().new_var().type_var().is_some());
@@ -496,7 +514,8 @@ fn test_solver_get_binding_after_bind() {
 fn test_inferrer_get_solver() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 获取求解器并使用
     let new_var = inferrer.solver().new_var();
