+++ src/frontend/typecheck/tests/infer.rs
@@ -4,8 +4,9 @@ use std::collections::HashMap;
 use crate::frontend::core::lexer::tokens::Literal;
 use crate::frontend::core::parser::ast::{BinOp, Block, Expr, Pattern, UnOp};
 use crate::frontend::typecheck::inference::ExprInferrer;
+use crate::frontend::typecheck::inference::ScopeManager;
+use crate::frontend::typecheck::inference::StatementChecker;
 use crate::frontend::typecheck::inference::patterns::PatternInferrer;
-use crate::frontend::typecheck::inference::statements::StmtInferrer;
 use crate::frontend::typecheck::*;
 use crate::frontend::typecheck::overload;
 use crate::util::span::Span;
@@ -15,7 +16,8 @@ use crate::util::span::Span;
 fn test_type_inferrer_new() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
     assert!(inferrer.solver().new_var().type_var().is_some());
 }
 
@@ -24,7 +26,8 @@ fn test_type_inferrer_new() {
 fn test_infer_literal_types() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 整数
     let int_expr = Expr::Lit(Literal::Int(42), Span::default());
@@ -57,7 +60,8 @@ fn test_infer_literal_types() {
 fn test_infer_unknown_variable() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let var_expr = Expr::Var("unknown_var".to_string(), Span::default());
     let result = inferrer.infer_expr(&var_expr);
@@ -69,7 +73,8 @@ fn test_infer_unknown_variable() {
 fn test_inferrer_add_get_var() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 添加变量
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -90,7 +95,8 @@ fn test_scope_management() {
     // 测试进入和退出作用域
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     inferrer.add_var("test".to_string(), PolyType::mono(MonoType::Int(64)));
 
@@ -111,7 +117,8 @@ fn test_scope_management() {
 fn test_infer_arithmetic_ops() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(1), Span::default());
     let right = Expr::Lit(Literal::Int(2), Span::default());
@@ -167,7 +174,8 @@ fn test_infer_arithmetic_ops() {
 fn test_infer_comparison_ops() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(1), Span::default());
     let right = Expr::Lit(Literal::Int(2), Span::default());
@@ -233,7 +241,8 @@ fn test_infer_comparison_ops() {
 fn test_infer_logical_ops() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Bool(true), Span::default());
     let right = Expr::Lit(Literal::Bool(false), Span::default());
@@ -264,7 +273,8 @@ fn test_infer_logical_ops() {
 fn test_infer_unary_ops() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let expr = Expr::Lit(Literal::Int(5), Span::default());
 
@@ -300,7 +310,8 @@ fn test_infer_unary_ops() {
 fn test_infer_tuple() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let elems = vec![
         Expr::Lit(Literal::Int(1), Span::default()),
@@ -323,7 +334,8 @@ fn test_infer_tuple() {
 fn test_infer_empty_list() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let list = Expr::List(vec![], Span::default());
 
@@ -342,7 +354,8 @@ fn test_infer_empty_list() {
 fn test_infer_dict() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let pairs = vec![
         (
@@ -370,7 +383,8 @@ fn test_infer_dict() {
 fn test_infer_empty_dict() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let dict = Expr::Dict(vec![], Span::default());
 
@@ -390,7 +404,8 @@ fn test_infer_empty_dict() {
 fn test_infer_assignment() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let left = Expr::Lit(Literal::Int(1), Span::default());
     let right = Expr::Lit(Literal::Int(2), Span::default());
@@ -411,7 +426,8 @@ fn test_infer_assignment() {
 fn test_infer_index() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 列表下标
     let list = Expr::List(
@@ -439,7 +455,8 @@ fn test_infer_index() {
 fn test_infer_tuple_static_index() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let tuple = Expr::Tuple(
         vec![
@@ -466,7 +483,8 @@ fn test_infer_tuple_static_index() {
 fn test_infer_tuple_out_of_bounds_index() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let tuple = Expr::Tuple(
         vec![
@@ -525,10 +543,11 @@ fn test_pattern_inferrer_fresh_type_vars() {
 
 #[test]
 fn test_stmt_inferrer_fresh_declaration_type_vars() {
-    let mut inferrer = StmtInferrer::new();
+    let mut solver = TypeConstraintSolver::new();
+    let mut checker = StatementChecker::new(&mut solver);
 
-    let t1 = inferrer.infer_declaration("a", None).unwrap();
-    let t2 = inferrer.infer_declaration("b", None).unwrap();
+    let t1 = checker.solver().new_var();
+    let t2 = checker.solver().new_var();
 
     match (t1, t2) {
         (MonoType::TypeVar(v1), MonoType::TypeVar(v2)) => assert_ne!(v1, v2),
@@ -613,7 +632,8 @@ fn test_infer_typed_pattern() {
 fn test_infer_return() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 有表达式的 return
     let ret_expr = Expr::Lit(Literal::Int(42), Span::default());
@@ -633,7 +653,8 @@ fn test_infer_return() {
 fn test_infer_break() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 无标签的 break
     let break_expr = Expr::Break(None, Span::default());
@@ -646,7 +667,8 @@ fn test_infer_break() {
 fn test_infer_continue() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 无标签的 continue
     let continue_expr = Expr::Continue(None, Span::default());
@@ -659,7 +681,8 @@ fn test_infer_continue() {
 fn test_infer_break_unknown_label() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let break_expr = Expr::Break(Some("unknown_label".to_string()), Span::default());
     let result = inferrer.infer_expr(&break_expr);
@@ -671,7 +694,8 @@ fn test_infer_break_unknown_label() {
 fn test_infer_continue_unknown_label() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let continue_expr = Expr::Continue(Some("unknown_label".to_string()), Span::default());
     let result = inferrer.infer_expr(&continue_expr);
@@ -683,7 +707,8 @@ fn test_infer_continue_unknown_label() {
 fn test_infer_cast() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let expr = Expr::Lit(Literal::Int(42), Span::default());
     let target_type = crate::frontend::core::parser::ast::Type::Int(64);
@@ -702,7 +727,8 @@ fn test_infer_cast() {
 fn test_infer_type_cast() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let expr = Expr::Lit(Literal::Int(42), Span::default());
     let cast = Expr::Cast {
@@ -720,7 +746,8 @@ fn test_infer_type_cast() {
 fn test_infer_function_call() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 定义一个函数变量
     let fn_type = MonoType::Fn {
@@ -751,7 +778,8 @@ fn test_infer_function_call() {
 fn test_infer_field_access() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 创建结构体类型
     let struct_type = MonoType::Struct(StructType {
@@ -783,7 +811,8 @@ fn test_infer_field_access() {
 fn test_infer_unknown_field() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let struct_type = MonoType::Struct(StructType {
         name: "Point".to_string(),
@@ -810,7 +839,8 @@ fn test_infer_unknown_field() {
 fn test_infer_field_access_on_non_struct() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
 
@@ -830,7 +860,8 @@ fn test_infer_field_access_on_non_struct() {
 fn test_infer_unsupported_op() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 函数定义在表达式上下文中是不支持的
     let fn_def = Expr::FnDef {
@@ -856,7 +887,8 @@ fn test_infer_unsupported_op() {
 fn test_infer_block() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let block = Block {
         stmts: vec![],
@@ -873,7 +905,8 @@ fn test_infer_block() {
 fn test_infer_empty_block() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let block = Block {
         stmts: vec![],
@@ -890,7 +923,8 @@ fn test_infer_empty_block() {
 fn test_infer_var_decl() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 使用 add_var 直接添加变量（模拟变量声明）
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -905,7 +939,8 @@ fn test_infer_var_decl() {
 fn test_infer_var_decl_with_annotation() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 使用 add_var 直接添加变量（模拟带类型的变量声明）
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -920,7 +955,8 @@ fn test_infer_var_decl_with_annotation() {
 fn test_infer_var_decl_no_initializer() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 使用 add_var 直接添加变量（模拟无初始化的变量声明）
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -935,7 +971,8 @@ fn test_infer_var_decl_no_initializer() {
 fn test_infer_if_expr() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let condition = Expr::Lit(Literal::Bool(true), Span::default());
     let then_body = Block {
@@ -966,7 +1003,8 @@ fn test_infer_if_expr() {
 fn test_infer_while_expr() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let condition = Expr::Lit(Literal::Bool(true), Span::default());
     let body = Block {
@@ -991,7 +1029,8 @@ fn test_infer_while_expr() {
 fn test_infer_for_loop() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let iterable = Expr::Lit(Literal::String("test".to_string()), Span::default());
     let body = Block {
@@ -1018,7 +1057,8 @@ fn test_infer_for_loop() {
 fn test_infer_neg() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let expr = Expr::Lit(Literal::Int(42), Span::default());
     let neg = Expr::UnOp {
@@ -1038,7 +1078,8 @@ fn test_infer_neg() {
 fn test_infer_fn_def() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     let fn_def = Expr::FnDef {
         name: "add".to_string(),
