+++ src/frontend/typecheck/tests/scope.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 use crate::frontend::typecheck::inference::ExprInferrer;
+use crate::frontend::typecheck::inference::ScopeManager;
 use crate::frontend::typecheck::overload;
 use crate::frontend::core::type_system::{MonoType, PolyType, TypeConstraintSolver};
 
@@ -7,7 +8,8 @@ use crate::frontend::core::type_system::{MonoType, PolyType, TypeConstraintSolve
 fn test_scope_shadowing() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // Global scope: x = Int
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -36,7 +38,8 @@ fn test_scope_shadowing() {
 fn test_nested_scopes() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // Global: a = Int
     inferrer.add_var("a".to_string(), PolyType::mono(MonoType::Int(64)));
