+++ src/frontend/typecheck/tests/shadowing.rs
@@ -5,9 +5,10 @@
 use std::collections::HashMap;
 use crate::frontend::core::parser::ast;
 use crate::frontend::typecheck::inference::ExprInferrer;
+use crate::frontend::typecheck::inference::ScopeManager;
 use crate::frontend::typecheck::overload;
 use crate::frontend::core::type_system::{MonoType, PolyType, TypeConstraintSolver};
-use crate::frontend::typecheck::checking::BodyChecker;
+use crate::frontend::typecheck::inference::BodyChecker;
 use crate::frontend::typecheck::TypeChecker;
 use crate::util::span::{Position, Span};
 
@@ -474,7 +475,8 @@ fn test_assignment_in_same_scope_ok() {
 fn test_inferrer_try_add_var_shadowing() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     // 全局：x = Int
     inferrer.add_var("x".to_string(), PolyType::mono(MonoType::Int(64)));
@@ -497,7 +499,8 @@ fn test_inferrer_try_add_var_shadowing() {
 fn test_inferrer_scope_destroyed_on_exit() {
     let mut solver = TypeConstraintSolver::new();
     let overload_candidates: HashMap<String, Vec<overload::OverloadCandidate>> = HashMap::new();
-    let mut inferrer = ExprInferrer::new(&mut solver, &overload_candidates);
+    let mut scope = ScopeManager::new();
+    let mut inferrer = ExprInferrer::new(&mut scope, &mut solver, &overload_candidates);
 
     inferrer.enter_scope();
     inferrer.add_var("temp".to_string(), PolyType::mono(MonoType::Bool));
